third_party/WebKit/LayoutTests/fast/canvas/canvas-arc-negative-radius.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<p id="description"></p>
<div id="console"></div>

<script>
test(function(t) {
    var canvas = document.createElement("canvas");
    var ctx = canvas.getContext('2d');

    ctx.arc(10, 10, 5, 0, 1, false);
    ctx.arc(10, 10, 0, 0, 1, false);
    try {
      ctx.arc(10, 10, -5, 0, 1, false);
    } catch (ex) {
      assert_equals(ex.name, "IndexSizeError");
    }

    ctx.arcTo(10, 10, 20, 20, 5);
    ctx.arcTo(10, 10, 20, 20, 0);
    try {
      ctx.arcTo(10, 10, 20, 20, -5);
    } catch (ex) {
      assert_equals(ex.name, "IndexSizeError");
    }

    var path = new Path2D();

    path.arc(10, 10, 5, 0, 1, false);
    path.arc(10, 10, 0, 0, 1, false);
    try {
      path.arc(10, 10, -5, 0, 1, false);
    } catch (ex) {
      assert_equals(ex.name, "IndexSizeError");
    }

    path.arcTo(10, 10, 20, 20, 5);
    path.arcTo(10, 10, 20, 20, 0);
    try {
      path.arcTo(10, 10, 20, 20, -5);
    } catch (ex) {
      assert_equals(ex.name, "IndexSizeError");
    }
}, 'Tests CanvasPath arc and arcTo with negative radii.');
</script>
